+++ packages/view-main/src/components/apis/dialog/App.svelte
@@ -69,7 +69,7 @@
 <Modal
   bind:visible
   teleport="#root"
-  minheight="140px"
+  minheight="8.75rem"
   bgclose={false}
   closebtn={false}
   title={extensionName}