+++ packages/web-server/src/app/renderer/winMain/preload/index.ts
@@ -16,7 +16,7 @@ console.log('preload win main')
 export type ExposeFunctions = AnyListen.IPC.ClientIPCActions<IPCSocket>
 export type ExposeServerFunctions = Omit<
   AnyListen.IPC.ServerIPC,
-  'showOpenDialog' | 'showSaveDialog' | 'openDirInExplorer' | 'openDevTools'
+  'showOpenDialog' | 'showSaveDialog' | 'openDirInExplorer' | 'openDevTools' | 'getSystemFonts'
 >
 export type ClientCall = AnyListen.IPC.ClientIPC