+++ packages/desktop/publish/changeLog.md
@@ -4,6 +4,7 @@
 - Added a **"Show lyrics in status bar"** option (MacOS only), located in _Settings > Playback Settings_. Disabled by default; can be enabled via _Settings > Playback Settings > Show lyrics in status bar_.
 - Added a **"Show lyrics in title bar"** option in _Settings > Playback Settings_. Disabled by default; can be enabled via _Settings > Playback Settings > Show lyrics in title bar_.
 - Added a **"Show lyrics in media control bar"** option in _Settings > Playback Settings_. Disabled by default; can be enabled via _Settings > Playback Settings > Show lyrics in media control bar_.
+- Added a **Font Settings** section under _Settings > Basic Settings > Font Settings_.
 
 ### Improved
 
@@ -26,6 +27,7 @@
 - 新增 **「显示状态栏歌词」** 选项（仅限 MacOS 版本），位于 _设置 > 播放设置_，默认关闭，可在 _设置 > 播放设置 > 显示状态栏歌词_ 启用。
 - 新增 **「显示标题栏歌词」** 选项，位于 _设置 > 播放设置_，默认关闭，可在 _设置 > 播放设置 > 显示标题栏歌词_ 启用。
 - 新增 **「显示媒体控制栏歌词」** 选项，位于 _设置 > 播放设置_，默认关闭，可在 _设置 > 播放设置 > 显示媒体控制栏歌词_ 启用。
+- 新增 **字体设置**，位于 _设置 > 基本设置 > 字体设置_。
 
 ### 优化